Change class name and Action ID
