+++ src/vs/workbench/browser/parts/editor/editorActions.ts
@@ -519,6 +519,31 @@ export class CloseEditorAction extends Action {
 	}
 }
 
+export class RevertAndCloseEditorAction extends Action {
+
+	public static ID = 'workbench.action.revertAndCloseActiveEditor';
+	public static LABEL = nls.localize('revertAndCloseActiveEditor', "Revert and Close Editor");
+
+	constructor(
+		id: string,
+		label: string,
+		@IWorkbenchEditorService private editorService: IWorkbenchEditorService,
+	) {
+		super(id, label);
+	}
+
+	public run(): TPromise<any> {
+		const activeEditor = this.editorService.getActiveEditor();
+		if (activeEditor) {
+			return activeEditor.input.revert().then(ok =>
+				this.editorService.closeEditor(activeEditor.position, activeEditor.input)
+			);
+		}
+
+		return TPromise.as(false);
+	}
+}
+
 export class CloseLeftEditorsInGroupAction extends Action {
 
 	public static ID = 'workbench.action.closeEditorsToTheLeft';
